ICB: Split detection features & adapt to new plugins
